+++ AOS/app/src/main/java/com/hyundai/myexperience/data/NoticeRepository.kt
@@ -1,7 +1,7 @@
 package com.hyundai.myexperience.data
 
-import com.hyundai.myexperience.data.entity.NoticeDetailItem
-import com.hyundai.myexperience.data.entity.NoticesItem
+import com.hyundai.myexperience.data.entity.notice.NoticeDetailItem
+import com.hyundai.myexperience.data.entity.notice.NoticesItem
 import com.hyundai.myexperience.data.mapper.mapToNoticeList
 import com.hyundai.myexperience.data.mapper.mapToNoticesItem
 import com.hyundai.myexperience.data.remote.NoticeRemoteDataSource
