--- AOS/app/src/main/java/com/hyundai/myexperience/data/entity/User.kt
@@ -1,5 +1,0 @@
-package com.hyundai.myexperience.data.entity
-
-data class User(
-    val message: String
-)
