+++ AOS/app/src/main/java/com/hyundai/myexperience/data/entity/notice/NoticeList.kt
@@ -1,4 +1,4 @@
-package com.hyundai.myexperience.data.entity
+package com.hyundai.myexperience.data.entity.notice
 
 data class NoticeList (
     val notices: List<NoticesItem>
