+++ AOS/app/src/main/java/com/hyundai/myexperience/data/entity/user/User.kt
@@ -1,0 +1,5 @@
+package com.hyundai.myexperience.data.entity.user
+
+data class User(
+    val message: String
+)
