+++ AOS/app/src/main/java/com/hyundai/myexperience/data/mapper/user/LoginResponseMapper.kt
@@ -1,7 +1,7 @@
 package com.hyundai.myexperience.data.mapper.user
 
 import com.hyundai.myexperience.data.dto.user.SignResponse
-import com.hyundai.myexperience.data.entity.User
+import com.hyundai.myexperience.data.entity.user.User
 
 fun SignResponse.mapToUser(): User {
     return User(
