+++ AOS/app/src/main/java/com/hyundai/myexperience/ui/notice/NoticeDetailViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.hyundai.myexperience.data.NoticeRepository
-import com.hyundai.myexperience.data.entity.NoticeDetailItem
+import com.hyundai.myexperience.data.entity.notice.NoticeDetailItem
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
