+++ AOS/app/src/main/java/com/hyundai/myexperience/ui/notice/adapter/NoticeViewHolder.kt
@@ -2,7 +2,7 @@ package com.hyundai.myexperience.ui.notice.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import com.hyundai.myexperience.databinding.ItemNoticeBinding
-import com.hyundai.myexperience.data.entity.NoticesItem
+import com.hyundai.myexperience.data.entity.notice.NoticesItem
 
 class NoticeViewHolder(private val itemBinding: ItemNoticeBinding) : RecyclerView.ViewHolder(itemBinding.root) {
     fun bind(noticesItem: NoticesItem, itemClickListener: NoticesItemClickListener) {
