+++ AOS/app/src/main/java/com/hyundai/myexperience/ui/notice/adapter/NoticesItemClickListener.kt
@@ -1,6 +1,6 @@
 package com.hyundai.myexperience.ui.notice.adapter
 
-import com.hyundai.myexperience.data.entity.NoticesItem
+import com.hyundai.myexperience.data.entity.notice.NoticesItem
 
 interface NoticesItemClickListener {
     fun onItemClick(notice : NoticesItem)
